sdk-format +  TargetFrameworks>net45;netcoreapp3.1;net5.0-windows + folder structur + changelog
