+++ ChangeLog.md
@@ -1,0 +1,6 @@
+# 0.2
+- 2021-08-21
+- TargetFrameworks: net45;netcoreapp3.1;net5.0-windows
+
+# 0.1.0 [2018]
+- first version, not published
