--- KsWare.PrivilegedExecutor.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27130.2027
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KsWare.PrivilegedExecutor", "KsWare.PrivilegedExecutor.csproj", "{8B3FBD46-105F-414B-9C61-053C6DAD0833}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "Tests\Tests.csproj", "{A8F85D83-619A-4095-A1C1-C3085EB6C531}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{8B3FBD46-105F-414B-9C61-053C6DAD0833}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8B3FBD46-105F-414B-9C61-053C6DAD0833}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8B3FBD46-105F-414B-9C61-053C6DAD0833}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8B3FBD46-105F-414B-9C61-053C6DAD0833}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A8F85D83-619A-4095-A1C1-C3085EB6C531}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A8F85D83-619A-4095-A1C1-C3085EB6C531}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A8F85D83-619A-4095-A1C1-C3085EB6C531}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A8F85D83-619A-4095-A1C1-C3085EB6C531}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {62700295-8ABD-47B0-82EE-A32BC416E246}
-	EndGlobalSection
-EndGlobal
