--- Properties/AssemblyInfo.cs
@@ -1,27 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-
-// General Information about an assembly is controlled through the following
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-[assembly: AssemblyTitle("KsWare.PrivilegedExecutor")]
-[assembly: AssemblyDescription("A privileged executor.")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("KsWare")]
-[assembly: AssemblyProduct("KsWare.PrivilegedExecutor")]
-[assembly: AssemblyCopyright("Copyright © 2018 by KsWare. All rights reserved.")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-
-[assembly: ComVisible(false)]
-[assembly: Guid("8b3fbd46-105f-414b-9c61-053c6dad0833")]
-[assembly: AssemblyVersion("0.1.0")]
-
-[assembly: InternalsVisibleTo("KsWare.PrivilegedExecutor.Tests")]
-
-namespace KsWare.PrivilegedExecutor {
-
-	public static class AssemblyInfo { }
-
-}
