--- Tests/Properties/AssemblyInfo.cs
@@ -1,21 +1,0 @@
-using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-
-[assembly: AssemblyTitle("KsWare.PrivilegedExecutor.Tests")]
-[assembly: AssemblyDescription("KsWare.PrivilegedExecutor.Tests")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("KsWare")]
-[assembly: AssemblyProduct("KsWare.PrivilegedExecutor")]
-[assembly: AssemblyCopyright("Copyright © 2018 by KsWare. All rights reserved.")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-
-[assembly: ComVisible(false)]
-
-[assembly: Guid("a8f85d83-619a-4095-a1c1-c3085eb6c531")]
-
-// [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.0.0.0")]
-
