--- packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="KsWare.Console" version="0.1.0-CI00002" targetFramework="net45" />
-  <package id="KsWare.IO.NamedPipes" version="0.1.0-CI00001" targetFramework="net45" />
-  <package id="KsWare.NuGet.PackageBuilder" version="0.1.0-CI00012" targetFramework="net45" developmentDependency="true" />
-</packages>
