+++ src/KsWare.PrivilegedExecutor.Tests/Properties/AssemblyInfo.cs
@@ -1,0 +1,8 @@
+﻿using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+
+[assembly: ComVisible(false)]
+
+[assembly: Guid("a8f85d83-619a-4095-a1c1-c3085eb6c531")]
+
