+++ src/KsWare.PrivilegedExecutor.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31613.86
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "KsWare.PrivilegedExecutor", "KsWare.PrivilegedExecutor\KsWare.PrivilegedExecutor.csproj", "{8B3FBD46-105F-414B-9C61-053C6DAD0833}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "KsWare.PrivilegedExecutor.Tests", "KsWare.PrivilegedExecutor.Tests\KsWare.PrivilegedExecutor.Tests.csproj", "{A8F85D83-619A-4095-A1C1-C3085EB6C531}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{A3BFA5B5-1D42-4A25-8D99-63D103F05BC3}"
+	ProjectSection(SolutionItems) = preProject
+		..\ChangeLog.md = ..\ChangeLog.md
+		..\README.md = ..\README.md
+	EndProjectSection
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{8B3FBD46-105F-414B-9C61-053C6DAD0833}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8B3FBD46-105F-414B-9C61-053C6DAD0833}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8B3FBD46-105F-414B-9C61-053C6DAD0833}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8B3FBD46-105F-414B-9C61-053C6DAD0833}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A8F85D83-619A-4095-A1C1-C3085EB6C531}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A8F85D83-619A-4095-A1C1-C3085EB6C531}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A8F85D83-619A-4095-A1C1-C3085EB6C531}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A8F85D83-619A-4095-A1C1-C3085EB6C531}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {62700295-8ABD-47B0-82EE-A32BC416E246}
+	EndGlobalSection
+EndGlobal
