+++ src/KsWare.PrivilegedExecutor/KsWare.PrivilegedExecutor.csproj
@@ -1,0 +1,50 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFrameworks>net45;netcoreapp3.1;net5.0-windows</TargetFrameworks>
+    <AssemblyTitle>KsWare.PrivilegedExecutor</AssemblyTitle>
+    <Company>KsWare</Company>
+    <Product>KsWare.PrivilegedExecutor</Product>
+    <Description>A privileged executor.</Description>
+    <Copyright>Copyright © 2018-2021 by KsWare. All rights reserved.</Copyright>
+    <Version>0.2.0</Version>
+    <AssemblyVersion>0.2.0</AssemblyVersion>
+    <FileVersion>0.2.0</FileVersion>
+    <Authors>SchreinerK</Authors>
+    <PackageProjectUrl>https://github.com/KsWare/KsWare.PrivilegedExecutor</PackageProjectUrl>
+    <PackageReleaseNotes></PackageReleaseNotes>
+    <PackageTags>execute elevated access-rights</PackageTags>
+  </PropertyGroup>
+  <PropertyGroup>
+    <PostBuildEvent />
+    <RepositoryUrl>https://github.com/KsWare/KsWare.PrivilegedExecutor.git</RepositoryUrl>
+    <RepositoryType>git</RepositoryType>
+    <PackageLicenseExpression>LGPL-3.0-only</PackageLicenseExpression>
+    <PackageIcon>packageicon.png</PackageIcon>
+    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+    <SignAssembly>true</SignAssembly>
+    <AssemblyOriginatorKeyFile>Properties\KsWare.snk</AssemblyOriginatorKeyFile>
+    <DelaySign>false</DelaySign>
+  </PropertyGroup>
+  <ItemGroup>
+    <PackageReference Include="KsWare.Console" Version="0.2.0" />
+    <PackageReference Include="KsWare.IO.NamedPipes" Version="0.2.0" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Remove="KsWare.PrivilegedExecutor.csproj.DotSettings" />
+    <None Include="..\..\packageicon.png">
+      <Pack>True</Pack>
+      <PackagePath></PackagePath>
+    </None>
+  </ItemGroup>
+  <ItemGroup Condition="'$(TargetFramework)' == 'netcoreapp3.1'">
+    <PackageReference Include="System.Security.Principal.Windows">
+      <Version>5.0.0</Version>
+    </PackageReference>
+  </ItemGroup>
+  <ItemGroup Condition="'$(TargetFramework)' == 'net5.0-windows'">
+    <PackageReference Include="System.Security.Principal.Windows">
+      <Version>5.0.0</Version>
+    </PackageReference>
+  </ItemGroup>
+</Project>
