+++ src/KsWare.PrivilegedExecutor/Properties/AssemblyInfo.cs
@@ -1,0 +1,14 @@
+﻿using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+
+[assembly: ComVisible(false)]
+[assembly: Guid("8b3fbd46-105f-414b-9c61-053c6dad0833")]
+
+[assembly: InternalsVisibleTo("KsWare.PrivilegedExecutor.Tests, PublicKey=0024000004800000940000000602000000240000525341310004000001000100C19B2E9069F6DFE3EA3857FC4F953C75C7BE0CB68C753999F472B78592FC320D80362F4E386A15D30D6A0D1EE8CC544094E5010B290309C0BFFA0A410F188B7F7C4A1800BD759C69416AB5DB5E3940842C9A5675EDB832B9FE25A9468E42ADBF4A2883F0AA2A28C39324971AF439439767633A27206E0673956CF8C2428856B2")]
+
+namespace KsWare.PrivilegedExecutor {
+
+	public static class AssemblyInfo { }
+
+}
