+++ src/KsWare.PrivilegedExecutor/desktop.ini
@@ -1,0 +1,2 @@
+[.ShellClassInfo]
+IconResource=D:\Data\Image Library\Icons\github.ico,0
